some changes
- add spinlock to ring buffer writing to avoid race with reading
- fix some typos which caused memory error in random parts of kernel (I spent 3 hours to find why my machine drops iin 'general protection failure' in 2 minutes after module loading and kmsg has no mentions of it in error trace...)

+++ main.c
@@ -10,7 +10,6 @@ struct ring_buffer *rbuf;
 struct kprobe **kp;
 
 /* for poll */
-//DEFINE_SPINLOCK(lock);
 DECLARE_WAIT_QUEUE_HEAD(wait_queue);
 bool polled = false;
 
@@ -83,7 +82,7 @@ static char *tracer_devnode(const struct device *dev, umode_t *mode) {
     return NULL;
 }
 
-static int __init my_kprobe_init(void) {
+static long __init my_kprobe_init(void) {
     int ret, i;
 
     rbuf = kmalloc(sizeof(struct ring_buffer), GFP_KERNEL);
